+++ tests/acceptance/setup-steps.R
@@ -22,7 +22,7 @@ cucumber::when("I run mutation tests with", function(code, context) {
   withr::with_output_sink(new = nullfile(), {
     withr::with_dir(context$dir, {
       suppressPackageStartupMessages(
-        context$score <- eval(parse(text = code))
+        context$score <- as.numeric(eval(parse(text = code)))
       )
     })
   })